requirements.txt
google-cloud-bigquery
google-auth
google-api-python-client
tenacity
functions-framework